[add] Timeline for ball movement

+++ src/resources/DimensionBundle.java
@@ -9,7 +9,7 @@ public class DimensionBundle extends ListResourceBundle {
     }
 
     static final Object[][] contents = {
-            {"SCREEN_HEIGHT", new Integer(900)},
+            {"SCREEN_HEIGHT", new Integer(800)},
             {"SCREEN_WIDTH", new Integer(500)},
             {"Ball_radius", new Double(10.0)},
             {"Jump_velocity", new Double(40.0)},
